backend/requirements.txt
Django
djangorestframework
mysqlclient
django-cors-headers
python-dateutil
